+++ CHANGELOG.md
@@ -19,3 +19,12 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 ### Added
 
 - Added emergency backup feature
+
+## 1.2.0 - 2020-04-26
+
+### Added
+
+- File pruning
+- "Latest" badge on utilties page to easily identify the most recent synced file
+- A "show more" dropdown when there are > 3 synced files
+- A "hide" setting to remove either databases or volumes from the utilities panel